Day 07 image scrambler
Made a script that converts a loaded image into an array of bytes, I wanted to shuffle the array and make a new image out of it to see some random glitch stuff. Only got as far to make it load an image into a array of bytes and then save a new image to drive from that array. Scrambler part will have to come later.