Day 14 String splitter
Asked Tropical Trevor for a challenge.
It was :
given an array of strings with names for each string see if there is a first and last name in the string print the initial of the first name and then the last name (Like Mark Smith becomes M Smith).
Ignore empty names or names without last name