Day 31 Machine learning agents Unity 1
After 2 days I finally have a fully working setup with Unity,tensorflow and the machine learning agents package. The agent(cube) has to find the target location(sphere) as quickly as possible.
Saving the whole unity file isn't possible with github at this time so these are the basic scripts with the agent settings and the training data settings.

**Thoughts:** I was away for over half a week and I needed a break of the java gui stuff after realizing I just may as well build it from scratch after working myself into a few corners :P
With the machine learning setup it's been a very difficult thing to get working. Its all still very much in experimental stages with th ML agents stuff. The differences in previous versions, lack of up to date documentation, using so many different pieces of software together. But finally its working !